Implemented PrioityQueue for soring people on Bus
The passngers on the bus should be sorted by their stopping location so we don't have to go through the everyone when unloading. So person now implements comparator.
I completely reworked location to be comparable and have a priority, but I only realized later that it would only work for a single bus on the route and that the priority would be relative to the bus that was asking. Gotta fix that.
I don't have an actual simulation running and that needs to be a priority.