Added Stop, Finished up Classes
I added a Bus Stop Class that acts as a queue and is different than Queue which is for events.
I realized that as long as people don't need to go to a stop before the current stop, priority can be a constant. I initialized all the stops and the locations in the simulartor file and gave each location a priority.
I split up the arrival process into dropoff/pickup methods. cuz if you are at the end of a loop and going to take a break you don't want to pick people up.
There is a lot of commented out code in Bus becuase I wanted to be able to set the starting location at different places.
I think all the classes are done. Propably ready to commit to master.